+++ examples/sampling/graphbolt/link_prediction.py
@@ -177,7 +177,7 @@ def create_dataloader(args, graph, features, itemset, is_train=True):
     # [Role]:
     # Convert a mini-batch to dgl-minibatch.
     ############################################################################
-    datapipe = gb.to_dgl_minibatch()
+    datapipe = gb.to_dgl()
 
     ############################################################################
     # [Input]:
